+++ src/app-components/app/RayLink/index.tsx
@@ -16,41 +16,41 @@ export default defineComponent({
   name: 'RayLink',
   setup() {
     const avatarOptions: AvatarOptions[] = [
-      {
-        key: 'yunhome',
-        src: 'https://yunkuangao.me/',
-        tooltip: '云之家',
-        icon: 'https://r2chevereto.yka.moe/avatar.jpeg',
-      },
-      {
-        key: 'yun-cloud-images',
-        src: 'https://yunkuangao.com/',
-        tooltip: '云图床',
-        icon: 'https://r2chevereto.yka.moe/avatar.jpeg',
-      },
+      // {
+      //   key: 'yunhome',
+      //   src: 'https://yunkuangao.me/',
+      //   tooltip: '云之家',
+      //   icon: 'https://r2chevereto.yka.moe/avatar.jpeg',
+      // },
+      // {
+      //   key: 'yun-cloud-images',
+      //   src: 'https://yunkuangao.com/',
+      //   tooltip: '云图床',
+      //   icon: 'https://r2chevereto.yka.moe/avatar.jpeg',
+      // },
       {
         key: 'ray-js-note',
         src: 'https://note.youdao.com/s/ObWEe2BB',
         tooltip: 'Ray的前端学习笔记',
-        icon: 'https://r2chevereto.yka.moe/longmao.navigator.png',
+        icon: 'https://avatars.githubusercontent.com/u/51957438?v=4',
       },
       {
         key: 'ray-js-cover',
         src: 'https://note.youdao.com/s/IC8xKPdB',
         tooltip: 'Ray的面试题总结',
-        icon: 'https://r2chevereto.yka.moe/longmao.navigator.png',
+        icon: 'https://avatars.githubusercontent.com/u/51957438?v=4',
       },
       {
         key: 'ray-template-doc',
         src: 'https://xiaodaigua-ray.github.io/ray-template-doc/',
         tooltip: 'Ray Template Doc',
-        icon: 'https://r2chevereto.yka.moe/longmao.navigator.png',
+        icon: 'https://avatars.githubusercontent.com/u/51957438?v=4',
       },
       {
         key: 'ray-template-doc-out',
         src: 'https://ray-template.yunkuangao.com/',
         tooltip: 'Ray Template Doc (国内地址)',
-        icon: 'https://r2chevereto.yka.moe/longmao.navigator.png',
+        icon: 'https://avatars.githubusercontent.com/u/51957438?v=4',
       },
     ]