+++ src/directives/modules/copy/index.ts
@@ -19,29 +19,31 @@ import ClipboardJS from 'clipboard'
 import type { CopyElement } from './type'
 import type { CustomDirectiveFC } from '@/directives/type'
 
+const createClipboard = (el: CopyElement, value: string) => {
+  const clipboard = new ClipboardJS(el, {
+    text: () => value,
+  })
+
+  el.$$clipboard = clipboard
+
+  el.$$clipboard.on('success', () => {
+    window.$message.success('复制成功')
+  })
+  el.$$clipboard.on('error', () => {
+    window.$message.error('复制失败')
+  })
+}
+
 const copyDirective: CustomDirectiveFC<CopyElement, string> = () => {
   return {
     mounted: (el, { value }) => {
-      const clipboard = new ClipboardJS(el, {
-        text: () => String(value),
-      })
-
-      clipboard.on('success', () => {
-        window.$message.success('复制成功')
-      })
-      clipboard.on('error', () => {
-        window.$message.error('复制失败')
-      })
-
-      el.$$clipboard = clipboard
+      createClipboard(el, value)
     },
     updated: (el, { value, oldValue }) => {
       if (value !== oldValue) {
         el.$$clipboard?.destroy()
 
-        el.$$clipboard = new ClipboardJS(el, {
-          text: () => String(value),
-        })
+        createClipboard(el, value)
       }
     },
     beforeUnmount: (el: CopyElement) => {